Breaks out config.
