+++ assistants/project-assistant/assistant/utils.py
@@ -6,11 +6,32 @@ codebase, helping to reduce code duplication and maintain consistency.
 """
 
 import logging
+import pathlib
 from typing import Optional, Tuple
 
 from semantic_workbench_assistant.assistant_app import ConversationContext
 
 logger = logging.getLogger(__name__)
+
+
+def load_text_include(filename) -> str:
+    """
+    Helper for loading an include from a text file.
+    
+    Args:
+        filename: The name of the text file to load from the text_includes directory
+        
+    Returns:
+        The content of the text file
+    """
+    # Get directory relative to this module
+    directory = pathlib.Path(__file__).parent
+    
+    # Get the file path for the prompt file
+    file_path = directory / "text_includes" / filename
+    
+    # Read the prompt from the file
+    return file_path.read_text()
 
 
 async def get_current_user(context: ConversationContext) -> Tuple[Optional[str], Optional[str]]:
